ec2-batch-processor/local-test/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
paddlepaddle>=2.6.0
paddleocr>=2.7.0
insightface>=0.7.3
onnxruntime>=1.16.0
boto3>=1.34.0
psycopg[binary]>=3.1.0
pgvector>=0.2.0
python-dotenv>=1.0.0
